[REFACTOR] Local Module 분리 원래 있던 곳은 Binds 로 바꿀거임
